Passing most tests in browser.
I'm using node-webcrypto-ossl
as a drop-in replacement for
the WebCrypto API.

All tests seem to pass, except
of course the web workers.

I am still not exactly sure how
web workers are used in this project,
but it could be worth taking on an
additional dep to make that test pass...
if it's easy.

In either case, I still need an explanation
for how that works in the JS version -
when and how it's called, and what
future plans are.